feat: Query system index without cluster name
We require system index query to happen at the local cluster.
Currently, OpenSearch does not support cross cluster system index
query. Thus, mapping of a remote index is unavailable.

Therefore, we require the local cluster to have the system index
of the remote cluster index.

The full "cluster:index" name is still
used to query OpenSearch for datarows, as CCS is natively supported.

Signed-off-by: Sean Kao <seankao@amazon.com>